added folders for components
Removed some files as well.
