+++ client/src/App.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
-import Navbar from './components/Navbar'
-import Landing from './components/Landing'
-import Login from './components/Login'
-import Register from './components/Register'
-import Profile from './components/Profile'
+import Navbar from './components/Navbar/Navbar'
+import Login from './components/Login/Login'
+import Register from './components/Register/Register'
+import Profile from './components/Profile/Profile'
 import './App.css';
 import Nav from './components/Nav/Nav.js'
 import Header from './components/Header/Header.js'
@@ -56,7 +55,6 @@ class App extends Component {
       <Router>
         <div className="App">
           <Navbar />
-          <Route exact path="/" component={Landing} />
           <div className="container">
             <Route exact path="/register" component={Register} />
             <Route exact path="/login" component={Login} />
