--- client/src/components/Landing.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react'
-
-class Landing extends Component {
-    render () {
-        return (
-            <div className="container">
-                <div className="jumbotron mt-5">
-                    <div className="col-sm-8 mx-auto">
-                        <h1 className="text-center">WELCOME</h1>
-                    </div>
-                </div>
-            </div>
-        )
-    }
-}
-
-export default Landing
