+++ client/src/components/Login/Login.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react'
-import { register } from './UserFunctions'
+import { login } from '../UserFunctions/UserFunctions'
 
-class Register extends Component {
+class Login extends Component {
     constructor() {
         super()
         this.state = {
-            name: '',
             email: '',
             password: ''
         }
@@ -22,13 +21,14 @@ class Register extends Component {
         e.preventDefault()
 
         const user = {
-            name: this.state.name,
             email: this.state.email,
             password: this.state.password
         }
 
-        register(user).then(res => {
-            this.props.history.push(`/login`)
+        login(user).then(res => {
+            if (res) {
+                this.props.history.push(`/profile`)
+            }
         })
     }
 
@@ -39,16 +39,6 @@ class Register extends Component {
                     <div className="col-md-6 mt-5 mx-auto">
                         <form noValidate onSubmit={this.onSubmit}>
                             <h1 className="h3 mb-3 font-weight-normal">Please sign in</h1>
-                            <div className="form-group">
-                                <label htmlFor="name">Name:</label>
-                                <input type="text"
-                                    className="form-control"
-                                    name="name"
-                                    placeholder="Enter Name"
-                                    value={this.state.name}
-                                    onChange={this.onChange}
-                                />
-                            </div>
                             <div className="form-group">
                                 <label htmlFor="email">Email Address</label>
                                 <input type="email"
@@ -71,7 +61,7 @@ class Register extends Component {
                             </div>
                             <button type="submit"
                                 className="btn btn-lg btn-primary btn-block">
-                                Register
+                                Sign in
                             </button>
                         </form>
                     </div>
@@ -81,4 +71,4 @@ class Register extends Component {
     }
 }
 
-export default Register
+export default Login
