+++ client/src/components/Register/Register.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react'
-import { login } from './UserFunctions'
+import { register } from '../UserFunctions/UserFunctions'
 
-class Login extends Component {
+class Register extends Component {
     constructor() {
         super()
         this.state = {
+            name: '',
             email: '',
             password: ''
         }
@@ -21,14 +22,13 @@ class Login extends Component {
         e.preventDefault()
 
         const user = {
+            name: this.state.name,
             email: this.state.email,
             password: this.state.password
         }
 
-        login(user).then(res => {
-            if (res) {
-                this.props.history.push(`/profile`)
-            }
+        register(user).then(res => {
+            this.props.history.push(`/login`)
         })
     }
 
@@ -39,6 +39,16 @@ class Login extends Component {
                     <div className="col-md-6 mt-5 mx-auto">
                         <form noValidate onSubmit={this.onSubmit}>
                             <h1 className="h3 mb-3 font-weight-normal">Please sign in</h1>
+                            <div className="form-group">
+                                <label htmlFor="name">Name:</label>
+                                <input type="text"
+                                    className="form-control"
+                                    name="name"
+                                    placeholder="Enter Name"
+                                    value={this.state.name}
+                                    onChange={this.onChange}
+                                />
+                            </div>
                             <div className="form-group">
                                 <label htmlFor="email">Email Address</label>
                                 <input type="email"
@@ -61,7 +71,7 @@ class Login extends Component {
                             </div>
                             <button type="submit"
                                 className="btn btn-lg btn-primary btn-block">
-                                Sign in
+                                Register
                             </button>
                         </form>
                     </div>
@@ -71,4 +81,4 @@ class Login extends Component {
     }
 }
 
-export default Login
+export default Register
